# schema-openapi/src/main/resources/openapi/openapi.yaml
# Tried pinning this to 3.0.3 to fix the Product1 duplication below (see ProductPage.yaml)
# on the theory that 3.1's dereferencer was the buggy party. It wasn't an improvement: 3.0's
# resolver doesn't chase a $ref found *inside* an externally-$ref'd fragment at all (a ref like
# ProductPage.yaml's `content.items: $ref: ./Product.yaml`, one level removed from openapi.yaml)
# - it silently gives up and types the field as `Object` instead of failing loudly. That's a
# worse outcome than a cosmetic duplicate class, so this stays on 3.1 despite the duplicate.
openapi: 3.1.0
info:
  title: Product Catalog API
  description: >
    Contract-first REST API for the product catalog. Every request/response schema is
    a components/schemas/<Name> entry that itself just $refs a standalone JSON Schema
    fragment under ./schemas/ - the fragment file is the actual source of truth; the
    components/schemas indirection exists only so OpenAPI Generator recognizes each one
    as a reusable named model instead of inlining a fresh synthetic type per operation.
    Operations must reference schemas via "#/components/schemas/<Name>", never directly
    via the fragment file path. The same fragments are used to generate Java types once
    (via OpenAPI Generator, in the `schema-openapi` Gradle module) and reused, by
    type-mapping in the `schema-graphql` module, in the equivalent GraphQL contract
    (../graphql/schema.graphqls, in that other module). Do not inline schemas here;
    add/edit a fragment file instead so REST and GraphQL stay backed by the same
    generated types.
  version: "1.0.0"
servers:
  - url: /api
tags:
  - name: Products
    description: CRUD and search operations over the product catalog.
paths:
  /products:
    get:
      operationId: listProducts
      tags: [Products]
      summary: List products (paged).
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: createProduct
      tags: [Products]
      summary: Create a new product.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: The created product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
  /products/search:
    get:
      operationId: searchProducts
      tags: [Products]
      summary: Search products by name, category and/or price range.
      description: >
        The one search endpoint. Deliberately takes scalar filter parameters (rather than
        a dedicated "criteria" object) so the same filter shape can be expressed
        identically as plain GraphQL field arguments - see `searchProducts` in
        schema.graphqls.
      parameters:
        - name: name
          in: query
          required: false
          description: Case-insensitive substring match on product name.
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: Exact match on category.
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: A page of matching products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /products/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getProductById
      tags: [Products]
      summary: Get a single product by id.
      responses:
        "200":
          description: The product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateProduct
      tags: [Products]
      summary: Update an existing product.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: The updated product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: deleteProduct
      tags: [Products]
      summary: Delete a product.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
    # Each entry below is the ONLY thing that makes OpenAPI Generator treat this as a
    # reusable named model instead of inlining a fresh synthetic type per operation. The
    # fragment file it points to remains the actual source of truth for the shape.
    Product:
      $ref: "./schemas/Product.yaml"
    ProductInput:
      $ref: "./schemas/ProductInput.yaml"
    ProductPage:
      $ref: "./schemas/ProductPage.yaml"
    ErrorResponse:
      $ref: "./schemas/ErrorResponse.yaml"
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
  responses:
    NotFound:
      description: No product exists with the given id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"